Moved contact listing table related functions to utils
